# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yalla Wassel delivery management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: fromZoneId
          in: query
          schema:
            type: integer
        - name: toZoneId
          in: query
          schema:
            type: integer
        - name: driverId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithDetails"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
        "404":
          description: Not found

    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"

    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Cancel/delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders/{id}/assign:
    patch:
      operationId: assignOrderDriver
      tags: [orders]
      summary: Assign a driver to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderAssign"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/{id}/eta:
    patch:
      operationId: setOrderEta
      tags: [orders]
      summary: Driver sets expected delivery time (ETA)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderEtaInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/track/{token}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by public token (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
        "404":
          description: Order not found

  # ── Ratings ────────────────────────────────────────────────────────────────────
  /ratings:
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings
      parameters:
        - name: driverId
          in: query
          schema:
            type: integer
        - name: orderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"

    post:
      operationId: createRating
      tags: [ratings]
      summary: Submit a rating for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Created rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"

  # ── Notifications ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OrderWithDetails:
      type: object
      required: [id, orderId, fromBusiness, fromAddress, fromZoneId, toCustomerName, toPhone, toAddress, toZoneId, priority, status, createdAt, trackingToken]
      properties:
        id:
          type: integer
        orderId:
          type: string
        fromBusiness:
          type: string
        fromAddress:
          type: string
        fromZoneId:
          type: integer
        fromZoneName:
          type: ["string", "null"]
        toCustomerName:
          type: string
        toPhone:
          type: string
        toAddress:
          type: string
        toZoneId:
          type: integer
        toZoneName:
          type: ["string", "null"]
        priority:
          type: string
          enum: [urgent, normal]
        status:
          type: string
          enum: [pending, assigned, picked, delivered, cancelled]
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        driverPhone:
          type: ["string", "null"]
        driverAvgRating:
          type: ["number", "null"]
        customerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        assignedAt:
          type: ["string", "null"]
        pickedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        eta:
          type: ["string", "null"]
        trackingToken:
          type: string
    OrderTracking:
      type: object
      required: [orderId, status, fromBusiness, toCustomerName, toAddress, toZoneName, priority, createdAt, trackingToken]
      properties:
        orderId:
          type: string
        status:
          type: string
          enum: [pending, assigned, picked, delivered, cancelled]
        fromBusiness:
          type: string
        toCustomerName:
          type: string
        toAddress:
          type: string
        toZoneName:
          type: ["string", "null"]
        priority:
          type: string
        driverName:
          type: ["string", "null"]
        driverAvgRating:
          type: ["number", "null"]
        createdAt:
          type: string
        assignedAt:
          type: ["string", "null"]
        pickedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        eta:
          type: ["string", "null"]
        trackingToken:
          type: string
        hasRating:
          type: boolean
        estimatedMinutes:
          type: ["integer", "null"]
...
    OrderInput:
      type: object
      required: [fromBusiness, fromAddress, fromZoneId, toCustomerName, toPhone, toAddress, toZoneId, priority]
      properties:
        fromBusiness:
          type: string
        fromAddress:
          type: string
        fromZoneId:
          type: integer
        toCustomerName:
          type: string
        toPhone:
          type: string
        toAddress:
          type: string
        toZoneId:
          type: integer
        priority:
          type: string
          enum: [urgent, normal]
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        fromBusiness:
          type: string
        fromAddress:
          type: string
        fromZoneId:
          type: integer
        toCustomerName:
          type: string
        toPhone:
          type: string
        toAddress:
          type: string
        toZoneId:
          type: integer
        priority:
          type: string
          enum: [urgent, normal]
        notes:
          type: string
    OrderAssign:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, assigned, picked, delivered, cancelled]
    OrderEtaInput:
      type: object
      required: [minutesFromNow]
      properties:
        minutesFromNow:
          type: integer
          minimum: 5
          maximum: 180

    # ── Rating schemas ─────────────────────────────────────────────────────────
    Rating:
      type: object
      required: [id, orderId, driverId, customerName, stars, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        driverId:
          type: integer
        customerName:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    RatingInput:
      type: object
      required: [orderId, driverId, customerName, stars]
      properties:
        orderId:
          type: integer
        driverId:
          type: integer
        customerName:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ── Notification schemas ──────────────────────────────────────────────────
...